prov/efa: move AV releted functions and structs to efa_av.h
Moved following structs:

    efa_ah
    efa_conn
    efa_av_entry
    efa_cur_reverse_av_key
    efa_cur_reverse_av
    efa_prv_reverse_av_key
    efa_prv_reverse_av

and following functions's declaration

    efa_av_open()
    efa_av_insert_one()
    efa_av_addr_to_conn()
    efa_av_reverse_lookup_rdm()
    efa_av_reverse_lookup_dgram()

from efa.h to efa_av.h because they are av related.

deleted the function rxr_ep_av() because it is not used.
