+++ prov/efa/src/dgram/efa_rma.c
@@ -37,6 +37,7 @@
 #include <ofi_iov.h>
 #include "efa_dgram.h"
 #include "efa.h"
+#include "efa_av.h"
 
 
 /*
