+++ prov/efa/src/efa_av.c
@@ -39,6 +39,7 @@
 #include <ofi_enosys.h>
 
 #include "efa.h"
+#include "efa_av.h"
 #include "rdm/rxr.h"
 #include "rdm/rxr_pkt_type_base.h"
